+++ src/index.js
@@ -1,1 +1,11 @@
-import './index.css';
+import './styles/main.css';
+import '@fortawesome/fontawesome-free/js/all.js';
+import { retriveAllData, countShows } from './modules/dom-utils.js';
+import setPopup from './modules/popUp_element.js';
+
+const main = document.querySelector('.main');
+
+window.onload = () => retriveAllData().then(() => {
+  countShows();
+});
+setPopup(main);
