+++ src/modules/constant_for_myproject.js
@@ -8,5 +8,5 @@ const tvWrapper = document.getElementById('tv-container');
 const INV_COMMENT_URL = `https://us-central1-involvement-api.cloudfunctions.net/capstoneApi/apps/${APP_ID}/comments`;
 
 export {
-    tvWrapper, API_URL, INV_LIKE_URL, INV_COMMENT_URL,
+  tvWrapper, API_URL, INV_LIKE_URL, INV_COMMENT_URL,
 };
