+++ src/template.html
@@ -1,16 +1,13 @@
 <!DOCTYPE html>
 <html lang="en">
+  <head>
+    <meta charset="UTF-8" />
+    <meta http-equiv="X-UA-Compatible" content="IE=edge" />
+    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
+    <title>Awesome tv</title>
+  </head>
 
-<head>
-  <meta charset="UTF-8" />
-  <meta http-equiv="X-UA-Compatible" content="IE=edge" />
-  <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-  <title>Webpack Setup</title>
-</head>
-
-<body>
-  <main>
-    <h1>Awsome TV</h1>
+  <body>
     <header>
       <a href="#" class="logo">Awesome tv</a>
       <nav class="nav">
@@ -26,10 +23,8 @@
     <main class="main">
       <section id="tv-container"></section>
     </main>
-</body>
-</main>
-<footer>
-  Created by Tetteh Kodjo-Sarso & Misikir Delelew &copy; All Rights Reserved
-</footer>
-
+    <footer>
+      Created by Tom Milanesi & Worash Bocherugn &copy; All Rights Reserved
+    </footer>
+  </body>
 </html>
